# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BarbrGo - Barber Booking Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /bookings:
    get:
      operationId: getBookings
      tags: [bookings]
      summary: List all bookings (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, completed, cancelled]
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
...
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: Get all users (admin)
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [customer, barber, admin]
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersResponse"
  /admin/bookings:
    get:
      operationId: getAdminBookings
      tags: [admin]
      summary: Get all bookings (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: All bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminBookingsResponse"
  /admin/revenue:
    get:
      operationId: getAdminRevenue
      tags: [admin]
      summary: Get revenue analytics
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, year]
      responses:
        "200":
          description: Revenue analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueAnalytics"

  # Platform summary dashboard
...
components:
  schemas:
...
    RegisterUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [customer, barber]
      required:
        - name
        - email
        - password
        - role
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
        - message
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [customer, barber, admin]
        avatar:
          type: string
        loyaltyPoints:
          type: integer
        isVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - loyaltyPoints
        - isVerified
        - createdAt

    # Barbers
...
    Service:
      type: object
      properties:
        id:
          type: integer
        barberId:
          type: integer
        serviceTypeId:
          type: integer
        name:
          type: string
        price:
          type: number
        durationMinutes:
          type: integer
        description:
          type: string
        isActive:
          type: boolean
      required:
        - id
        - barberId
        - name
        - price
        - durationMinutes
        - isActive

    # Bookings
    CreateBookingRequest:
      type: object
      properties:
        userId:
          type: integer
        barberId:
          type: integer
        serviceIds:
          type: array
          items:
            type: integer
        date:
          type: string
        startTime:
          type: string
        isHomeService:
          type: boolean
        address:
          type: string
        paymentMethod:
          type: string
          enum: [online, cash]
        couponCode:
          type: string
        notes:
          type: string
      required:
        - userId
        - barberId
        - serviceIds
        - date
        - startTime
        - paymentMethod
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        barberId:
          type: integer
        barberName:
          type: string
        barberAvatar:
          type: string
        userName:
          type: string
        services:
          type: array
          items:
            $ref: "#/components/schemas/Service"
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        totalDuration:
          type: integer
        totalAmount:
          type: number
        discountAmount:
          type: number
        finalAmount:
          type: number
        isHomeService:
          type: boolean
        address:
          type: string
        paymentMethod:
          type: string
          enum: [online, cash]
        paymentStatus:
          type: string
          enum: [pending, paid, refunded]
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        couponCode:
          type: string
        notes:
          type: string
        cancelReason:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - barberId
        - barberName
        - userName
        - services
        - date
        - startTime
        - endTime
        - totalDuration
        - totalAmount
        - discountAmount
        - finalAmount
        - isHomeService
        - paymentMethod
        - paymentStatus
        - status
        - createdAt

    # Reviews
...
    AdminUsersResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
      required:
        - users
        - total
    AdminBookingsResponse:
      type: object
      properties:
        bookings:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
        total:
          type: integer
      required:
        - bookings
        - total
    RevenueAnalytics:
      type: object
      properties:
        period:
          type: string
        total:
          type: number
        breakdown:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              revenue:
                type: number
              bookings:
                type: integer
            required:
              - label
              - revenue
              - bookings
      required:
        - period
        - total
        - breakdown

    # Dashboard
...