# lib/api-zod/src/generated/types/createBookingRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BarbrGo - Barber Booking Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateBookingRequestPaymentMethod } from "./createBookingRequestPaymentMethod";

export interface CreateBookingRequest {
  userId: number;
  barberId: number;
  serviceIds: number[];
  date: string;
  startTime: string;
  isHomeService?: boolean;
  address?: string;
  paymentMethod: CreateBookingRequestPaymentMethod;
  couponCode?: string;
  notes?: string;
}