# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BarbrGo - Barber Booking Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  barberId: number;
  serviceTypeId?: number;
  name: string;
  price: number;
  durationMinutes: number;
  description?: string;
  isActive: boolean;
}